+++ Release Notes.md
@@ -25,6 +25,7 @@ SOFTWARE.
 
 * arm_cmsis: updated latest chips references for STMicro
 * arm_cmsis: platform CMSIS includes & HAL config included from there
+* arm_cmsis: removed including HAL config (already included from CMSIS includes)
 * arm_hal_peripheral: added to include used peripherals
 * arm_hal_peripheral: reworked all peripheral includes
 * arm_chip_xx: Families divided into files with respective fam name